refactor(Calendar): 훅, 유틸함수 이동
